Use types of fixed size.

+++ languages/c++/benchmarks/map.cpp
@@ -7,7 +7,7 @@
 //
 // Container type to test.
 //
-using Container = std::map<long, long>;
+using Container = std::map<int32_t, int32_t>;
 
 //
 // Parameters of Lorenz attractor.
